[RISCV] got riscv qemu tracing up and running; did profile run on iam

+++ RISCV/src/org/specs/Riscv/instruction/RiscvInstruction.java
@@ -58,7 +58,7 @@ public class RiscvInstruction extends AInstruction {
      */
     private RiscvInstruction(String address, String instruction, InstructionData idata,
             RiscvAsmFieldData fieldData, InstructionProperties props) {
-        super(Integer.parseInt(address, 16), instruction, idata, props);
+        super(Long.parseLong(address, 16), instruction, idata, props);
         this.fieldData = fieldData;
     }
